Changed host label abbreviation algorithm so that only one affix is selected: whichever is more common.
This is to fix bug where a hostname may appear to not
be abbreviated, when in fact it consists of nothing but both the common suffix
and common prefix. By only permitting one or the other of those to exist at any
one time, the appropriate affix will always be selected.